* förberett funktion för att uppdatera bokning från server * ändrat text för widget * förbättrat information från server

+++ server/src/main/java/se/acrend/christopher/server/service/impl/BillingServiceImpl.java
@@ -87,7 +87,7 @@ public class BillingServiceImpl {
       }
       if (type == ProductType.Count) {
 
-        long currentCount = (Long) subscription.getProperty("notificationCount");
+        int currentCount = (Integer) subscription.getProperty("notificationCount");
 
         int count = Integer.parseInt(value);
